refactor: split citum-io lib.rs into formats/
Extract all format-specific private helpers from lib.rs into
focused submodules under src/formats/:

- biblatex.rs — BibLaTeX bibliography loading (~30 lines)
- native.rs   — Citum YAML/JSON/CBOR serde helpers (~392 lines)
- csl_json.rs — CSL-JSON load/write and name conversion (~114)
- ris.rs      — RIS parse helpers (~129 lines)
- output.rs   — BibLaTeX and RIS text renderers with shared
  contributor_name_strings helper (~94 lines)

lib.rs becomes a thin facade: public API, RefsFormat /
LoadedBibliography types, and dispatch arms into formats/.
Non-test code reduced from ~1622 to ~320 lines (-80%).

DRY: removed duplicate load_hybrid_json_reference; unified
into parse_hybrid_json_reference.

Performance: parse_hybrid_json_reference uses &str comparison
for class/type routing, avoiding two String allocations/call.

Idiom: validate_compound_sets import moved to module level;
if/else None/Some → .then_some() in the merge path.

Docs: added # Errors sections to all public Result-returning
functions to satisfy clippy::missing_errors_doc.

Checks off item 8 (citum-io/src/lib.rs) in bean csl26-rfct.
